Support for UTF-16 file paths and names on Windows.
This allows for full unicode character support for file paths, file names, and asset names on Windows.

If both the _WIN32 and _NIRACLIENT_UTF16_PATHS_AND_NAMES preprocessor flags are defined,
the name and path parameters of the array of NiraAssetFile structs passed to niraUploadAsset
can be wchar_t strings rather than char strings. The assetName parameter passed to
niraUploadAsset can also be a wchar_t string.

Wide versions of string parameters to niraclient-c's other functions will probably be added later,
but this is a good start to ensure all characters are supported for filen ames and asset names on
Windows.